requirements.txt
Django==4.0.6
pillow==11.0.0
django-admin-sortable2==2.1.10
